Display usage logs in reverse order (most recent first):
This is a much more natural order for display (most recent at the top),
but it means we're sorting twice.

That should be a negligible penalty though, and we'd much rather incur
a performance hit when displaying usage logs than when writing them.

(Standard lists are optimized for appending to the end - inserts at
the beginning don't perform well. There's deques for that, but the
persistent implementation in `persistent.ring._DequeRing` isn't
considered public).