Switch to `ipaddress` module instead of `py2-ipaddress` for IP range parsing,
and fix unicode handling:

We previously included both the py2-ipaddress and ipaddress packages as
dependencies, py2-ipaddress directly from ftw.tokenauth, and ipaddress as
an indirect dependency of the cryptography module.

Because they both occupy the same module namespace, any one of them was
picked at random, depending on module load order. Because the ipaddress
module absolutely requires IP specifications to be passed in as unicode,
this lead to sporadic failures.

Since the ipaddress module is already an (indirect) dependency, and it
is more actively maintained, we drop the dependency on py2-ipaddress and
use the ipaddress module instead.

This means we need to ensure we always give it unicode though, so some
tests that directly instantiate ip_networks need to be fixed, and the
functions for parsing IP range specifications need to ensure they always
convert their input to unicode if necessary.